Tutorial code breaking
While following this documentation for learning purpose, i fall into the issue where `mongoDs.automigrate` was raising an exception (`automigrate is not a function of undefined`). Bit of investigation leads to following solution:

  var mongoDs = app.dataSources['mongo-connector'];

Here, `mongo-connector` is the name of datasource which i created while defining the datasource. It could be mapped to datasource.json as well. 
Note: Their should be a generic method which should return a connected datasource. This will add some level of independence.