Updated create-angularjs-client.md (#298)
* Updated create-angularjs-client.md

In my experience with just one installation, "lb-ng" will not be installed by following the tutorial as written.  I added instructions to install the "apiconnect" NPM module which provides "lb-ng".  I also found that the folder containing "lb-ng" was not added to the path so I added a step to include the necessary folder in the path environment variable.

* Changed package to loopback-sdk-angular-cli