Update Embedded-models-and-relations.md (#345)
* Update Embedded-models-and-relations.md

Added missing "model" property to Link relations as this makes the example clearer.

* Update Embedded-models-and-relations.md

Also added an example of the embedded model for embedsMany as it's not specified that the embedded model needs a string id set with defaultFn of uuid to function when sending multiple embedded modes. Otherwise the request fails with unique id error.